requirements.txt
openai>=1.66.0
matrix-nio>=0.25.2
python-dotenv>=1.0.1
pydantic>=2.10.6
pytest>=8.3.4
pytest-asyncio>=0.25.3
